v.1.15.0.0

+++ engine/src/wolf.content_pipeline/collada/c_animation.h
@@ -272,7 +272,7 @@ namespace wolf
 						{
 							if (pTime >= this->times[i] && pTime < this->times[i + 1])
 							{
-								_index = i;
+								_index = (unsigned int)i;
 								break;
 							}
 						}
